+++ views/feature-toggle-status.ejs
@@ -1,51 +1,47 @@
 <link rel="stylesheet" href="/css/wide.css">
-<style>
-  table {
-    width: 80%;
-    margin: 20px auto;
-    border-collapse: collapse;
-  }
-
-  table th,
-  table td {
-    padding: 10px;
-    text-align: center;
-    border: 1px solid var(--border-color);
-  }
-
-  table th {
-    background-color: var(--background-color);
-  }
-
-  .enabled {
-    background-color: #41ff6f94;
-  }
-
-  .disabled {
-    background-color: #ff031894;
-  }
-
-  .status-label {
-    font-weight: bold;
-  }
-</style>
 <h1>Feature Toggle Status</h1>
+<div class="features">
+  <% 
+  // Group the features based on removing the first part of the name
+  const groupedFeatures = {};
+  
+  for (const [feature, { enabled, maintenance, coming_soon }] of Object.entries(featureToggles)) {
+    const groupName = feature.split(' ').slice(1).join(' ');
+    if (!groupedFeatures[groupName]) {
+      groupedFeatures[groupName] = [];
+    }
+    groupedFeatures[groupName].push([feature, { enabled, maintenance, coming_soon }]);
+  }
 
-<table>
-  <thead>
-    <tr>
-      <th>Feature Name</th>
-      <th>Status</th>
-    </tr>
-  </thead>
-  <tbody>
-    <% for (const [feature, enabled] of Object.entries(featureToggles)) { %>
-    <tr>
-      <td><%= feature %></td>
-      <td class="<%= enabled ? 'enabled' : 'disabled' %>">
-        <span class="status-label"><%= enabled ? 'Enabled' : 'Disabled' %></span>
-      </td>
-    </tr>
-    <% } %>
-  </tbody>
-</table>
+  // Sort the groups alphabetically
+  const sortedGroupedFeatures = Object.keys(groupedFeatures).sort().reduce((acc, key) => {
+    acc[key] = groupedFeatures[key];
+    return acc;
+  }, {});
+%>
+
+  <% // Loop through the sorted groups and display them
+  for (const [groupName, features] of Object.entries(sortedGroupedFeatures)) { %>
+  <h3><%= groupName.charAt(0).toUpperCase() + groupName.slice(1) %> Features</h3>
+  <table>
+    <thead>
+      <tr>
+        <th>Feature Name</th>
+        <th>Status</th>
+      </tr>
+    </thead>
+    <tbody>
+      <% features.forEach(([feature, { enabled, maintenance, coming_soon }]) => { %>
+      <tr>
+        <td><%= feature %></td>
+        <td class="<%= (coming_soon ? 'coming_soon ' : (maintenance ? 'maintenance ' : (enabled ? 'enabled' : 'disabled')))%>">
+          <span class="status-label">
+            <% if (coming_soon) { %>Coming Soon<% } else if (maintenance) { %>Maintenance<% } else { %><%= enabled ? 'Enabled' : 'Disabled' %><% } %>
+          </span>
+        </td>
+      </tr>
+      <% }) %>
+    </tbody>
+  </table>
+  <% } %>
+</div>
